рефакторинг функции render
